# swagger.yaml
openapi: 3.0.0
info:
  title: Academic Workload Analyzer API
  description: RESTful API documentation for managing courses, tasks, and tracking individual student workloads.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server

# Apply global JWT security verification to data endpoints by default
security:
  - BearerAuth: []

# Define the groups (tags)
tags:
  - name: Auth
    description: Public endpoints for user registration and authentication
  - name: Dashboard
    description: Statistics and metrics for the main dashboard
  - name: Courses
    description: Endpoints for managing academic courses
  - name: Tasks
    description: Endpoints for managing assignments and tasks

paths:
  # ================= AUTHENTICATION (PUBLIC) =================
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Register a new student account
      security: [] # Overrides global lock to make this route public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: "ahmed123"
                password:
                  type: string
                  example: "securepass123"
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Username already exists or missing data elements

  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Authenticate user and receive session JWT token
      security: [] # Overrides global lock to make this route public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: "ahmed123"
                password:
                  type: string
                  example: "securepass123"
      responses:
        '200':
          description: Authentication successful, access token returned
        '401':
          description: Invalid username or password entry

  # ================= DASHBOARD (PROTECTED) =================
  /api/dashboard/stats:
    get:
      tags:
        - Dashboard
      summary: Retrieve isolated workload dashboard statistics
      responses:
        '200':
          description: Successfully retrieved statistics for current user session
        '401':
          description: Unauthorized access. Valid JWT token missing

  # ================= COURSES (PROTECTED) =================
  /api/courses:
    get:
      tags:
        - Courses
      summary: Get all courses belonging to logged-in user
      responses:
        '200':
          description: A list of courses owned by the user
    post:
      tags:
        - Courses
      summary: Create a new course tied to current user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: "System Analysis and Design"
                instructor:
                  type: string
                  example: "Dr. Ahmed"
      responses:
        '201':
          description: Course created successfully

  /api/courses/{id}:
    put:
      tags:
        - Courses
      summary: Update an existing course
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Numeric ID of the course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                instructor:
                  type: string
      responses:
        '200':
          description: Course updated successfully
        '404':
          description: Course not found or user unauthorized to modify it
    delete:
      tags:
        - Courses
      summary: Delete a course and its associated tasks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Course deleted successfully
        '404':
          description: Course not found or user unauthorized to delete it

  # ================= TASKS (PROTECTED) =================
  /api/tasks:
    get:
      tags:
        - Tasks
      summary: Get all tasks belonging to logged-in user
      responses:
        '200':
          description: A list of user tasks with associated course names
    post:
      tags:
        - Tasks
      summary: Create a new task tied to current user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  example: "Finish SAD Project"
                description:
                  type: string
                  example: "Complete the code and documentation"
                status:
                  type: string
                  enum: [pending, in-progress, completed]
                  example: "pending"
                deadline:
                  type: string
                  format: date
                  example: "2026-05-21"
                course_id:
                  type: integer
                  example: 1
      responses:
        '201':
          description: Task created successfully

  /api/tasks/{id}:
    put:
      tags:
        - Tasks
      summary: Update a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
                deadline:
                  type: string
                  format: date
                course_id:
                  type: integer
      responses:
        '200':
          description: Task updated successfully
        '404':
          description: Task not found or user unauthorized to modify it
    delete:
      tags:
        - Tasks
      summary: Delete a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task deleted successfully
        '404':
          description: Task not found or user unauthorized to delete it

# Cryptographic and protocol layer parameter handling definition definitions
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Paste your valid session JWT token (obtained via login endpoint) here to safely query protected user operations.